neo: move "invalid entities" retry out of client into neo command
The retry only makes sense for `pulumi neo`, where the stack is a
convenience hint. The `--neo-task-on-failure` path uses the stack
link to scope the debug task, so retrying without it would silently
strip that context. Move the wrapper into pkg/cmd/pulumi/neo so
client.CreateNeoTask stays neutral.
