+++ changelog/pending/20260511--cli-neo--retry-task-creation-without-stack-entity-on-invalid-entities.yaml
@@ -1,4 +1,4 @@
 changes:
 - type: fix
   scope: cli/neo
-  description: Retry `pulumi neo` task creation without the stack entity when the backend rejects it with an "invalid entities" error
+  description: Retry `pulumi neo` task creation without the attached stack when the backend rejects it with an "invalid entities" error
